🚀 feat: interactive; titiler geojson
