+++ frontEnd/public/app.conf.json
@@ -3,7 +3,7 @@
     "back_app": "/api",
     "back_app2": "/api2",
     "back_websocket": "",
-    "minioIpAndPort": "http://192.168.1.101:30900",
+    "minioIpAndPort": "http://192.168.1.110:30900",
     "titiler": "/tiler",
 
     "ndviProjectId": "PRJyxITmCVLn3PMoaeca",
