+++ inDB/json/log.txt
@@ -1,24 +1,16 @@
 ========== [Landsat8_OLI] ==========
 [Scene 1] E:\Data\0708\LC08_L2SP_121035_20250413_20250416_02_T1\LC08_L2SP_121035_20250413_20250416_02_T1.tiff
-[FAILED] Scene 1, E:\Data\0708\LC08_L2SP_121035_20250413_20250416_02_T1\LC08_L2SP_121035_20250413_20250416_02_T1.tiff
-[ERROR] insert_to_db() failed
-[WARNING] Data may have been partially written, please check manually.
+[NOT STARTED] Scene 1, E:\Data\0708\LC08_L2SP_121035_20250413_20250416_02_T1\LC08_L2SP_121035_20250413_20250416_02_T1.tiff, program exited unexpectedly
 [Scene 2] E:\Data\0708\LC08_L2SP_122034_20250420_20250425_02_T1\LC08_L2SP_122034_20250420_20250425_02_T1.tiff
-[FAILED] Scene 2, E:\Data\0708\LC08_L2SP_122034_20250420_20250425_02_T1\LC08_L2SP_122034_20250420_20250425_02_T1.tiff
-[ERROR] insert_to_db() failed
-[WARNING] Data may have been partially written, please check manually.
+[NOT STARTED] Scene 2, E:\Data\0708\LC08_L2SP_122034_20250420_20250425_02_T1\LC08_L2SP_122034_20250420_20250425_02_T1.tiff, program exited unexpectedly
 [Scene 3] E:\Data\0708\LC08_L2SP_122035_20250420_20250425_02_T1\LC08_L2SP_122035_20250420_20250425_02_T1.tiff
-[FAILED] Scene 3, E:\Data\0708\LC08_L2SP_122035_20250420_20250425_02_T1\LC08_L2SP_122035_20250420_20250425_02_T1.tiff
-[ERROR] insert_to_db() failed
-[WARNING] Data may have been partially written, please check manually.
+[NOT STARTED] Scene 3, E:\Data\0708\LC08_L2SP_122035_20250420_20250425_02_T1\LC08_L2SP_122035_20250420_20250425_02_T1.tiff, program exited unexpectedly
 ========== Batch Insert End for this product ==========
 
 ========== [Landsat9_OLI] ==========
 [Scene 1] E:\Data\0708\LC09_L2SP_122034_20250412_20250413_02_T1\LC09_L2SP_122034_20250412_20250413_02_T1.tiff
-[FAILED] Scene 1, E:\Data\0708\LC09_L2SP_122034_20250412_20250413_02_T1\LC09_L2SP_122034_20250412_20250413_02_T1.tiff
-[ERROR] insert_to_db() failed
-[WARNING] Data may have been partially written, please check manually.
+[NOT STARTED] Scene 1, E:\Data\0708\LC09_L2SP_122034_20250412_20250413_02_T1\LC09_L2SP_122034_20250412_20250413_02_T1.tiff, program exited unexpectedly
 [Scene 2] E:\Data\0708\LC09_L2SP_123034_20250419_20250420_02_T1\LC09_L2SP_123034_20250419_20250420_02_T1.tiff
-[SUCCESS] Scene 2, E:\Data\0708\LC09_L2SP_123034_20250419_20250420_02_T1\LC09_L2SP_123034_20250419_20250420_02_T1.tiff
+[NOT STARTED] Scene 2, E:\Data\0708\LC09_L2SP_123034_20250419_20250420_02_T1\LC09_L2SP_123034_20250419_20250420_02_T1.tiff, program exited unexpectedly
 ========== Batch Insert End for this product ==========
