+++ inDB/json/productDEMConfig.json
@@ -1,0 +1,13 @@
+{
+    "SCENE_PREFIX": "DEM",
+    "TEMP_OUTPUT_DIR": "D:\\code\\test",
+    "CUR_SENSOR_NAME": "DEM",
+    "CUR_PRODUCT_NAME": "DEM产品",
+    "CUR_RESOLUTION": "30m",
+    "CUR_PERIOD": "16d",
+    "TAGS": {
+        "source": "national",
+        "production": "light",
+        "category": "normal"
+    }
+}
