--- modelServer/dataProcessing/history/80cb3a17-2038-48f8-bde0-d3728f1ae7b8.json
@@ -1,10 +1,0 @@
-{
-  "STATUS": "COMPLETE",
-  "RESULT": {
-    "bucket": "temp-files",
-    "object_path": "mosaicjson/207bda0c-392c-4068-8d29-17bcf75d3de9.json"
-  },
-  "TIME": "2025-06-20 23:26:55",
-  "ID": "80cb3a17-2038-48f8-bde0-d3728f1ae7b8",
-  "MD5": "d9927d3ea2e33efe693eacce68deb1cd"
-}
